requirements.txt
pyorbital
pyephem
geopy
skyfield
pyfiglet
